+++ History.md
@@ -1,4 +1,10 @@
 
+0.1.0 / 2011-12-18
+==================
+
+  * change Parse(args []string, defaults []string, description string, flags int) -> ParseCommandLine(description string, flags int) ... read args and env by os.Args and os.Environ
+  * use path/filepath.Base function
+
 0.0.2 / 2011-12-13
 ==================